% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCCA_functions.R
\name{addPattern}
\alias{addPattern}
\title{cluster weight vectors by assigning them to one pattern}
\usage{
addPattern(List.Pattern, corr, xyz.vector, z.vector)
}
\arguments{
\item{List.Pattern}{}

\item{corr}{}

\item{xyz.vector}{}

\item{z.vector}{}
}
\description{
cluster weight vectors by assigning them to one pattern in list 
if mean cluster vector is similar to new z.vector;
make new cluster if not similar to any existing cluster
count number of vectors associated to one cluster, 
save correlation for cluster 
cluster items:
item [[1]]: correlation
item [[2]]: combined vector of x, y, and z-weight
item [[3]]: counter for number of vectors in cluster
}
