% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCCA_functions.R
\name{comparePattern}
\alias{comparePattern}
\title{compare patterns of two eigenvectors}
\usage{
comparePattern(x, y, dm = c("manhattan", "euklidean"))
}
\arguments{
\item{x, y}{vectors to compare}

\item{dm}{distance measure, default is manhattan}
}
\value{
manhattan distance by default; other distances possible
}
\description{
return manhattan distance by default; other distances possible
}
