% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.best.lambdas.R
\name{get.best.lambdas}
\alias{get.best.lambdas}
\title{Calculate supervised CCA on three data matrices}
\usage{
get.best.lambdas(X, Z, lambda.grid, n.r = 10, max.counter.test = 10)
}
\arguments{
\item{X, Z}{biological data sets, Z is design data set}

\item{lambda.grid}{set of lambda combinations}

\item{n.r}{number of resampling steps}

\item{max.counter.test}{number of random start vectors for iteration (inner loop)}
}
\value{
list with three components:
      best.lambdas: optimal lambdas for the data sets
}
\description{
Calculate correlation coefficients for sparse CCA for all combinations 
of lambdas (using grid search).
}
\details{
Adequate numbers for the grid search depend on the data set size. 
As a starting point 0.0 is recommended, this is equivalent to no sparsity. 
Depending on the number of features, 
for about 1000 features, as ending point 0.25 should be high enough, 
for about 10 features it is more like 2. 
For the step size 0.01 is a good number for sets with 1000 features, 0.1 for 10 features. 
For a good resampling quality set n.r to about ten.
The higher the number of random start vectors, the better. 
Depending on the instability of the data set between 5 and 20 should be good.

As design data set Z is small, lambdas are higher and bigger steps can be used 
for data set with about 1000 features, lambda.end=0.3 should be high enough; 
use higher end values for smaller data sets and lower for bigger data sets.
}
\examples{
TRUE
}
\author{
Anjana Bhat, Andrea Thum, Elena Parkhomenko
}
