% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCCA3.R
\name{getCCA3}
\alias{getCCA3}
\title{Calculate supervised CCA on three or more data matrices}
\usage{
getCCA3(
  X,
  Z,
  end,
  step,
  numCV = 10,
  n.r = 10,
  max.counter.test = 10,
  grid.search = "exhaustive",
  n.comb = 5000
)
}
\arguments{
\item{X}{list of biological data sets with dim(X1, X2,..): n rows/samples, p features}

\item{Z}{design data set with dim(Z): n rows/samples, r features; describes experimental design with binary vectors}

\item{end, step}{end and stepsize for lambda.(x,y,z) grid search.}

\item{numCV}{number of canonical variables which will be calculated}

\item{n.r}{number of resampling runs, 10 is okay}

\item{max.counter.test}{number of random start vectors for iteration (inner loop)}

\item{grid.search}{search strategy for lambda: exhaustive considers all possible lambda combinations, random considers sampled set of lambda combinations}

\item{n.comb}{number of lambda combinations to be sampled if random grid search is selected}
}
\value{
A list with the elements 
  cc3.weights.(xz): Matrices with weights of canonical variables for data sets X and Z respectively (columnwise)
  cc3.CV.(xz): matrices with canonical variables for data set X and Z respectively (columnwise)
  corr: matrix with correlation coefficient for pairwise (X-Z,Y-Z) and correlation of can. variable: mean(abs(corr(X-Z)),abs(corr(Y-Z))) (columnwise)
  lambda: best lambda.(xyz) for each data set and for each can. var. (columnwise)
  numCV: Number of canonical variables
}
\description{
getCCA3() starts the supervised sparse CCA
}
\details{
The getCCA3() function performs sparse 
canonical correlation analysis on three or more data matrices 
(n Biological data sets in X, and one design data set Z) with elastic net. 
Data sets in X are strongly regularized (ridge regression); Z if neccessary (det(Z)==0).

The getCCA3 first normalizes the data matrices (column-wise mean centering and unit variance), 
computes canonical variables, 
and updates the data matrices by removing the latent variables iteratively 
for a specified number of canonical variables (numCV). 
It returns a list containing matrices with weight vectors and canonical variables, 
correlation coefficients, and optimal sparsity parameters (lambda values) 
for each canonical variable.
}
\examples{
TRUE
}
\author{
Anjana Bhat, Andrea Thum
}
\keyword{CCA}
\keyword{sparse}
\keyword{supervised}
