% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCCA.R
\name{plotCCA}
\alias{plotCCA}
\title{Plot CCA results to PDF file(s)}
\usage{
plotCCA(CCA3, X, Experiments = NULL, filename = NULL, legend)
}
\arguments{
\item{CCA3}{list with getCCA3() information to plot}

\item{X}{original biological data matrices without subtraction of canonical variables, for plots of top features}

\item{Experiments}{Sample names for X-axis. If NULL, rownames from CCA3 are considered}

\item{filename}{If NULL, figures are plotted on screen, otherwise CCA-project name as base output filename (_CV_numcv.pdf and _Top_numcv.pdf will be appended)}

\item{legend}{names of data sets X, Z for legend, e.g. c("Genes", "Metabolites","Design")}
}
\description{
Plot Canonical Variables and top two features for each canonical variable
}
\examples{
TRUE
}
\author{
Andrea Thum
}
