% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.CCA.R
\name{save.CCA}
\alias{save.CCA}
\title{Save CCA3 results to (human readable) file}
\usage{
save.CCA(CCA3, filename, omic.names)
}
\arguments{
\item{CCA3}{list with getCCA3() information to save}

\item{filename}{CCA-project name as base output filename (_numcv.txt be appended)}

\item{omic.names}{names of biological datasets (X) to separate selected features list in text file

 @importFrom utils tail}
}
\description{
Store results of current canonical variable in a file. 
First row: information about lambdas and correlation coefficients. 
Next rows for all three data sets for each weight > 0 weight and feature name. 
Weights are sorted in decreasing order, so most important features are on top.
}
\examples{
TRUE
}
\author{
Andrea Thum
}
