% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.function3Z.R
\name{scca.function3Z}
\alias{scca.function3Z}
\title{Title scca.function3Z}
\usage{
scca.function3Z(
  XpZ,
  ZpX,
  dims,
  lambda.x,
  lambda.z,
  max.iter = 100,
  shift = 2,
  x.initial = list(),
  z.initial = NULL
)
}
\arguments{
\item{XpZ, ZpX}{For regularized pseudomatrix Xp of X: regularize(t(X)*X)) * t(X); then XpZ = Xp * Z is input. Z is the design data set.}

\item{dims}{number of features of all data sets including Z}

\item{lambda.x, lambda.z}{lasso parameters lambda for data sets. lambda = 0: no sparsity}

\item{max.iter}{max. number of iterations for power method}

\item{shift}{start lasso regularization after shift iteration steps}

\item{x.initial, z.initial}{initial vectors for iteration. If NULL: filled with random values}
}
\value{
a list with the components:
  (xz).new: eigenvectors, i.e. weight vectors for supervised sparse CCA for all data sets (X;Z) and given sparsity (lambdas)
}
\description{
Implements the modified iterative power method for three (regularized) data sets 
with lasso penalty for given lasso parameters lambda.(x,y,z).
Pseudomatrices are used as input for faster calculation.
This function is called by get.best.lambdas()
}
\note{
Test output variable n, if iteration did converge (n<max.iter)!
}
\examples{
TRUE
}
\author{
Andrea Thum, Elena Parkhomenko
}
\keyword{CCA}
\keyword{lasso}
